Move generic status into results namespace
Each concrete scanner should simply have a sub-namespace in which specific status class can live.
