+++ packages/Antivirus/src/Scanners/NullScanner.php
@@ -2,7 +2,7 @@
 
 namespace Aedart\Antivirus\Scanners;
 
-use Aedart\Antivirus\Scanners\Status\GenericStatus;
+use Aedart\Antivirus\Results\GenericStatus;
 use Aedart\Contracts\Antivirus\Results\ScanResult;
 use Aedart\Contracts\Streams\FileStream;
